requirements.txt
# Web framework
Flask==2.3.2
Werkzeug==2.3.7
click>=8.1.3
itsdangerous>=2.1.2
Jinja2>=3.1.2
blinker>=1.6.2

# Database
SQLAlchemy==2.0.23

# HTML parsing
beautifulsoup4==4.12.3
soupsieve==2.5

# HTTP requests
requests==2.31.0
urllib3==2.2.2
certifi>=2024.7.4
charset-normalizer>=3.3.2
idna>=3.7

# Other utils
typing_extensions>=4.12.2

python-dotenv==1.1.1
